Feature

+++ server/app/api/users.py
@@ -4,6 +4,7 @@ from api.dependencies import follower_service, get_user, user_service
 from config import setting
 from fastapi import APIRouter, Depends, Path
 from schemas.common import SuccessSchemaResponse
+from schemas.errors import ErrorSchemaResponse
 from schemas.followers import FollowerSchemaAddModel
 from schemas.users import UserSchemaResponse
 from services.followers import FollowerService
@@ -53,11 +54,11 @@ async def get_info_about_your_profile(
     return await get_user_info(user_id, user_service)
 
 
-@router.get("/{id}")
+@router.get("/{id}", responses={404: {"model": ErrorSchemaResponse}})
 async def get_info_about_other_profile_by_id(
     user_service: Annotated[UserService, Depends(user_service)],
     id: int = Path(
-        ..., title="id пользователя", description="id пользователя"
+        ..., title="id пользователя", description="id пользователя", gt=0
     ),
 ) -> UserSchemaResponse:
     """
@@ -69,24 +70,41 @@ async def get_info_about_other_profile_by_id(
     return await get_user_info(id, user_service)
 
 
-@router.post("/{id}/follow")
+@router.post("/{id}/follow", responses={404: {"model": ErrorSchemaResponse}})
 async def follow_user_by_id(
+    user_service: Annotated[UserService, Depends(user_service)],
     follower_service: Annotated[FollowerService, Depends(follower_service)],
     id: int = Path(
         ...,
         title="id подписываемого пользователя",
         description="id подписываемого пользователя",
+        gt=0,
     ),
     user_id=Depends(get_user),
 ) -> SuccessSchemaResponse:
     """
     Эндпоинт позволяет добавить пользователя в читаемые
     :param id: id пользователя добавляемого в читаемые
+    :param user_service: сервис работы с БД для пользователей
     :param follower_service: сервис работы с БД для
     добавления/удаления пользователей в читаемые
     :param user_id: id текущего пользователя
     :return:
     """
+    # делаем проверку на то что пользователь с таким id существует в БД
+    # если проверка не пройдена, то ничего не меняем в БД
+    # выводим отрицательный результат на фронтенд
+    user = await user_service.get_user_by_id(id)
+    if user is None:
+        raise ClientHTTPException(
+            status_code=404, detail=f"Пользователя с id={id} не найден."
+        )
+    # проверка на то что пользователь добавляет не сам себя
+    # если проверка не пройдена, то ничего не меняем в БД
+    # выводим положительный результат на фронтенд
+    if id == user_id:
+        return {"result": True}
+
     # проверяем добавлен ли уже пользователь в читаемые
     # если уже добавлен, то ничего не меняем в БД
     # выводим положительный результат на фронтенд
@@ -103,32 +121,39 @@ async def follow_user_by_id(
     return {"result": True}
 
 
-@router.delete("/{id}/follow")
+@router.delete("/{id}/follow", responses={404: {"model": ErrorSchemaResponse}})
 async def unfollow_user_by_id(
+    user_service: Annotated[UserService, Depends(user_service)],
     follower_service: Annotated[FollowerService, Depends(follower_service)],
     id: int = Path(
         ...,
         title="id подписанного пользователя",
         description="id подписанного пользователя",
+        gt=0,
     ),
     user_id=Depends(get_user),
 ):
     """
     Эндпоинт позволяет удалять пользователя из читаемых
     :param id: id пользователя удаляемого из читаемых
+    :param user_service: сервис работы с БД для пользователей
     :param follower_service: сервис работы с БД
     для добавления/удаления пользователей в читаемые
     :param user_id: id текущего пользователя
     :return:
     """
-    result = await follower_service.delete_following(
+    # делаем проверку на то что пользователь с таким id существует в БД
+    # если проверка не пройдена, то ничего не меняем в БД
+    # выводим отрицательный результат на фронтенд
+    user = await user_service.get_user_by_id(id)
+    if user is None:
+        raise ClientHTTPException(
+            status_code=404, detail=f"Пользователя с id={id} не найден."
+        )
+
+    # делаем запрос к БД для удаления пользователя из читаемых
+    await follower_service.delete_following(
         user_id_follower=user_id, user_id_following=id
     )
-    # делаем запрос к БД для удаления пользователя из читаемых
-    if not result:
-        raise ClientHTTPException(
-            status_code=404,
-            detail=f"Пользователь с id={user_id} отсутствует в "
-            f"списке читаемых для пользователя с id={id}",
-        )
-    return {"result": result}
+
+    return {"result": True}
